\name{REAT-package}
\alias{REAT-package}
\alias{REAT}
\docType{package}
\title{
Regional Economic Analysis Tools
}
\description{
This package contains a collection of models and analysis methods used in regional and urban economics and (quantitative) economic geography. The functions in this package can be divided in seven groups: (1) analysis of regional disparities and inequality (including \emph{Gini coefficient} and \emph{Lorenz curve}), (2) specialization of regions (including \emph{spatial Gini coefficient of regional specialization} and \emph{Krugman coefficient for regional specialization}), (3) spatial concentration of industries (including \emph{spatial Gini coefficient for industry concentration}), (4) regional growth (including traditional \emph{shift-share analysis}), (5) spatial interaction and accessibility models (including \emph{Huff model} and \emph{Hansen accessibility}), (6) proximity analysis (including calculation of \emph{distance matrices} and \emph{buffers}) and (7) additional tools for data preparation und visualization. The package also contains data examples. 
}

\author{
Thomas Wieland

Maintainer: Thomas Wieland \email{thomas.wieland.geo@googlemail.com}
}

\references{
Dinc, M. (2002): \dQuote{Regional and Local Economic Analysis Tools}. Prepared for the Public Finance, Decentralization and Poverty Reduction Program (World Bank Institute). Washington, D.C. : The World Bank. 

Doersam, P. (2004): \dQuote{Wirtschaftsstatistik anschaulich dargestellt}. Heidenau : PD-Verlag.

Farhauer, O./Kroell, A. (2013): \dQuote{Standorttheorien: Regional- und Stadtoekonomik in Theorie und Praxis}. Wiesbaden : Springer.

Guessefeldt, J. (1999): \dQuote{Regionalanalyse}. Muenchen : Oldenbourg.

Lessmann, C. (2005): \dQuote{Regionale Disparitaeten in Deutschland und ausgesuchten OECD-Staaten im Vergleich}. \emph{ifo Dresden berichtet}, \bold{3/2005}. 

Loeffler, G. (1998): \dQuote{Market areas - a methodological reflection on their boundaries}. In: \emph{GeoJournal}, \bold{45}, 4, p. 265-272.
}